# api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Jobjoo API Documentation"
host: "localhost:4000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /basic-info/countries:
    get:
      tags:
      - "basic-info"
      summary: "Get List Of All Countries"
      operationId: "getBasicinfoCountries"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
    post:
      tags:
      - "basic-info"
      summary: "Post Countries [Only Admins]"
      operationId: "postBasicinfoCountries"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 5"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
...
  /basic-info/languages:
    get:
      tags:
      - "basic-info"
      summary: "Get List Of All Languages"
      operationId: "getBasicinfoLanguages"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
    post:
      tags:
      - "basic-info"
      summary: "Post Language (Only Admins)"
      operationId: "postBasicinfoLanguages"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 6"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
...
  /basic-info/skills-category:
    get:
      tags:
      - "basic-info"
      summary: "Get List Of All Skill Categories"
      operationId: "getBasicinfoSkillscategory"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
    post:
      tags:
      - "basic-info"
      summary: "Post Skills Category [Only Admins]"
      operationId: "postBasicinfoSkillscategory"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 8"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
...
  /profile/profile-image:
    get:
      tags:
      - "profile"
      summary: "Get Profile Image"
      operationId: "getProfileProfileimage"
      parameters:
      - name: "responseType"
        in: "query"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Profile"
    post:
      tags:
      - "profile"
      summary: "Upload Profile Image"
      operationId: "postProfileProfileimage"
      consumes:
      - "multipart/form-data"
      parameters:
      - name: "data"
        in: "formData"
        description: "Select a file to upload"
        required: true
        type: "file"
        x-meta:
          swaggerType: "file"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Profile"
...
  /basic-info/cities/{stateid}:
    get:
      tags:
      - "basic-info"
      summary: "Get List Of All Cities"
      operationId: "getBasicinfoCitiesStateid"
      parameters:
      - name: "stateid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/skills/{categoryid}:
    get:
      tags:
      - "basic-info"
      summary: "Get List Of All Skills"
      operationId: "getBasicinfoSkillsCategoryid"
      parameters:
      - name: "categoryid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/states/{countryid}:
    get:
      tags:
      - "basic-info"
      summary: "Get List Of All States"
      operationId: "getBasicinfoStatesCountryid"
      parameters:
      - name: "countryid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
...
  /basic-info/cities:
    post:
      tags:
      - "basic-info"
      summary: "Post Cities [Only Admins]"
      operationId: "postBasicinfoCities"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 4"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/skills:
    post:
      tags:
      - "basic-info"
      summary: "Post Skill [Only Admins]"
      operationId: "postBasicinfoSkills"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 7"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/states:
    post:
      tags:
      - "basic-info"
      summary: "Post States [Only Admins]"
      operationId: "postBasicinfoStates"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 9"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
...
  /profile/recruiter-data:
    post:
      tags:
      - "profile"
      summary: "Store Recruiter Profile Data"
      operationId: "postProfileRecruiterdata"
      consumes:
      - "multipart/form-data"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 13"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Profile"
  /profile/user-data:
    post:
      tags:
      - "profile"
      summary: "Store Primary Profile Data"
      operationId: "postProfileUserdata"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 14"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Profile"
...
  /basic-info/category-name/{categoryid}:
    post:
      tags:
      - "basic-info"
      summary: "Get Specific Category Details [Only Admins]"
      operationId: "postBasicinfoCategorynameCategoryid"
      parameters:
      - name: "categoryid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/location-name/{id}:
    post:
      tags:
      - "basic-info"
      summary: "Get Location Details Based On Id"
      operationId: "postBasicinfoLocationnameId"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 29"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
...
  /basic-info/cities/{cityid}:
    delete:
      tags:
      - "basic-info"
      summary: "Delete Specific City [Only Admins]"
      operationId: "deleteBasicinfoCitiesCityid"
      parameters:
      - name: "cityid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/countries/{countryid}:
    delete:
      tags:
      - "basic-info"
      summary: "Delete Specific Country [Only Admins]"
      operationId: "deleteBasicinfoCountriesCountryid"
      parameters:
      - name: "countryid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/languages/{languageid}:
    delete:
      tags:
      - "basic-info"
      summary: "Delete Specific Language [Only Admins]"
      operationId: "deleteBasicinfoLanguagesLanguageid"
      parameters:
      - name: "languageid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/skills/{skillid}:
    delete:
      tags:
      - "basic-info"
      summary: "Delete Specific Skill [Only Admins]"
      operationId: "deleteBasicinfoSkillsSkillid"
      parameters:
      - name: "skillid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/skills-category/{categoryid}:
    delete:
      tags:
      - "basic-info"
      summary: "Delete Specific Category [Only Admins]"
      operationId: "deleteBasicinfoSkillscategoryCategoryid"
      parameters:
      - name: "categoryid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /basic-info/states/{stateid}:
    delete:
      tags:
      - "basic-info"
      summary: "Delete Specific State [Only Admins]"
      operationId: "deleteBasicinfoStatesStateid"
      parameters:
      - name: "stateid"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "BasicInfo"
  /profile/profile-type/{profile_type}:
    put:
      tags:
      - "profile"
      summary: "Set Profile Type (Recruiter / Recruitee)"
      operationId: "putProfileProfiletypeProfile_type"
      parameters:
      - name: "profile_type"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Profile"
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...